pyproject.toml
[project]
name = "sine"
version = "0.1.0"
description = "Sionna-based Network Emulation - Wireless network emulation using Sionna ray tracing and Containerlab"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Joshua" }
]
keywords = ["network", "emulation", "wireless", "sionna", "ray-tracing", "containerlab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Networking",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "httpx>=0.24",
    "docker>=6.0",
    "rich>=13.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
gpu = [
    "tensorflow==2.15.0",
    "sionna>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
sine = "sine.cli:main"

[project.urls]
Homepage = "https://github.com/joshua/sine"
Documentation = "https://github.com/joshua/sine#readme"
Repository = "https://github.com/joshua/sine"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sine"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
